// client/src/app/drawing-tools/drawing-tool-constants.ts
export const enum TOOL_SELECTED {
    pencil = 'pencil',
    paintBrush = 'paintBrush',
    rectangle = 'rectangle',
    ellipse = 'ellipse',
    pen = 'pen',
}

export const enum TEXTURES_ID {
    texture1 = 'f1',
    texture2 = 'f2',
    texture3 = 'f3',
    texture4 = 'f4',
    texture5 = 'f5',
}

export const enum drawingSpecifics {
    stroke = 'stroke',
    strokeWidth = 'stroke-width',
    fill = 'fill',
    elementID = 'elementID',
}

export const enum rectAttributes {
    id = 'id',
    x = 'x',
    y = 'y',
    height = 'height',
    width = 'width',
}

export const enum polygonAttributes {
    id = 'id',
    height = 'height',
    width = 'width',
    points = 'points',
    transform = 'transform',
}

export const enum ellipseAttributes {
    id = 'id',
    cx = 'cx',
    cy = 'cy',
    ry = 'ry',
    rx = 'rx',
}

export const enum stampAttributes {
    id = 'id',
    height = 'height',
    width = 'width',
    transform = 'transform',
    rotate = 'rotate',
    x = 'x',
    y = 'y',
}

export const enum EMOJIS {
    happy = 'happy',
    angry = 'angry',
    bored = 'bored',
    confused = 'confused',
    kissing = 'kissing',
}